requirements.txt
fastapi
uvicorn
python-dotenv
requests
gspread
google-auth
pydantic
